custom image name for common workflow
